+++ src/com/adverserealms/astar/core/CostTile.scala
@@ -30,5 +30,5 @@ trait CostTile {
   /**
    * Returns the cost for this tile
    */
-  def getCost() : Double
+  def getCost() : Float
 }
